+++ lib/expansion_tile_list.dart
@@ -1,3 +1,4 @@
+/// Library for ExpansionTileList and ExpansionTileExtension
 library expansion_tile_list;
 
 export 'src/expansion_tile_extension.dart';