+++ CHANGELOG.md
@@ -1,2 +1,10 @@
 ## 0.0.1
+
 - Expansion Tile List release
+
+## 0.1.0
+
+- Added `trailing` property which cascades to all the tiles
+- Added `trailingAnimation` property which is a `Animatable<double>`
+- Added `trailingAnimationBuilder` property to customize the trailing widget animation
+- Added `trailingAnimationEnabled` property which enabled/disabled trailing animation is true by default